First pass for PowerShell install and launch scripts.

+++ launch-script.ps1
@@ -1,14 +1,21 @@
-echo "Stopping Freelancer and FLServer"
+$freelancer = Get-Process freelancer -ErrorAction SilentlyContinue
+if ($freelancer){
+Write-Output "Stopping any instances of Freelancer.exe"
 Get-Process "freelancer" | Stop-Process
+}
+$flserver = Get-Process flserver -ErrorAction SilentlyContinue
+if ($flserver){
+Write-Output "Stopping any instances of flserver.exe"
 Get-Process "flserver" | Stop-Process
+}
 
-echo "Downloading latest update"
+Write-Output "Downloading latest update"
 #git checkout main
 $git =  "https://github.com/TheStarport/freelancer-bmod.git"
 git pull $git
 
-
 #echo "Starting the game"
 Start-Process -FilePath "EXE/Freelancer.exe" -ArgumentList "-s162.248.93.98:2302 -w"
 
-Read-Host -Prompt "Press Enter to close this console window"
+Remove-Variable freelancer
+Remove-Variable flserver
